+++ mux/src/functions.h
@@ -1,6 +1,6 @@
 // functions.h -- declarations for functions & function processing.
 //
-// $Id: functions.h,v 1.19 2006-01-07 21:53:26 sdennis Exp $
+// $Id: functions.h,v 1.20 2006-01-11 20:51:31 sdennis Exp $
 //
 
 #include "copyright.h"
@@ -75,7 +75,7 @@ bool delim_check
 void arr2list(char *arr[], int alen, char *list, char **bufc, SEP *psep);
 int list2arr(char *arr[], int maxlen, char *list, SEP *psep);
 char *trim_space_sep(char *str, SEP *psep);
-char *trim_space_sep_LEN(char *str, int nStr, SEP *psep, int *nTrim);
+char *trim_space_sep_LEN(char *str, size_t nStr, SEP *psep, size_t *nTrim);
 char *next_token(char *str, SEP *psep);
 char *split_token(char **sp, SEP *psep);
 int countwords(char *str, SEP *psep);
