+++ mux/src/help.h
@@ -1,14 +1,14 @@
 // help.h
 //
-// $Id: help.h,v 1.1 2003-01-22 19:58:25 sdennis Exp $
+// $Id: help.h,v 1.2 2006-01-11 20:51:31 sdennis Exp $
 //
 
 #define  TOPIC_NAME_LEN     30
 
 typedef struct
 {
-  long pos;         /* index into help file */
-  int len;          /* length of help entry */
+  size_t pos;         /* index into help file */
+  size_t len;         /* length of help entry */
   char topic[TOPIC_NAME_LEN + 1];   /* topic of help entry */
 } help_indx;
