+++ mux/src/look.cpp
@@ -1,6 +1,6 @@
 // look.cpp -- Commands which look at things.
 //
-// $Id: look.cpp,v 1.48 2006-01-11 04:19:53 jake Exp $
+// $Id: look.cpp,v 1.49 2006-01-11 20:51:31 sdennis Exp $
 //
 #include "copyright.h"
 #include "autoconf.h"
@@ -474,7 +474,7 @@ static void look_exits(dbref player, dbref loc, const char *exit_name)
         tPtr = FormatOutput;
 
         char *preserve[MAX_GLOBAL_REGS];
-        int preserve_len[MAX_GLOBAL_REGS];
+        size_t preserve_len[MAX_GLOBAL_REGS];
         save_and_clear_global_regs("look_exits_save", preserve, preserve_len);
 
         mux_exec(FormatOutput, &tPtr, loc, player, player,
@@ -650,7 +650,7 @@ static void look_contents(dbref player, dbref loc, const char *contents_name, in
             { VisibleObjectList, ContentsNameScratch };
 
         char *preserve[MAX_GLOBAL_REGS];
-        int preserve_len[MAX_GLOBAL_REGS];
+        size_t preserve_len[MAX_GLOBAL_REGS];
         save_and_clear_global_regs("look_contents_save", preserve, preserve_len);
 
         mux_exec(FormatOutput, &tPtr, loc, player, player,
@@ -1221,7 +1221,7 @@ void look_in(dbref player, dbref loc, int key)
         char *tPtr = FormatOutput;
 
         char *preserve[MAX_GLOBAL_REGS];
-        int preserve_len[MAX_GLOBAL_REGS];
+        size_t preserve_len[MAX_GLOBAL_REGS];
         save_and_clear_global_regs("look_in_save", preserve, preserve_len);
 
         mux_exec(FormatOutput, &tPtr, loc, player, player,
